backend/pyproject.toml
[project]
name = "driving-school-backend"
version = "0.1.0"
description = "Backend for Driving School Management System"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.16",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.1"
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "ISC", "G", "PIE", "T20", "PT", "Q", "RET", "SIM", "ARG", "PTH", "PGH", "PL", "TRY", "RUF"]
ignore = ["E501"]
fixable = ["ALL"]
unfixable = []
exclude = ["alembic", ".venv", "venv", "build", "dist"]
line-length = 120
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"alembic/**" = ["ALL"]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | alembic
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/alembic/*", "*/__init__.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

